+++ java/alix/lucene/search/ScorerOccs.java
@@ -17,7 +17,7 @@ public class ScorerOccs extends Scorer
   }
 
   @Override
-  public float score(final int occsMatch, final long docLen)
+  public double score(final int occsMatch, final long docLen)
   {
     return occsMatch;
   }
